Integrate time-based segmentation into dms-segments.py
Move the time-based boundary logic into a single via_time function in dms-segments.py behind a new --time-based-segments switch (default false, preserving existing behavior). Remove the standalone dms-segments-time-based.py and update the README to document the switch.
